# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{userId}/role": {
      "patch": {
        "operationId": "UsersController_updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user role",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/devices": {
      "post": {
        "operationId": "UsersController_registerDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDevice"
                }
              }
            }
          }
        },
        "summary": "Register a device for push notifications",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_getUserDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of user devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDevice"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all registered devices for user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/devices/{deviceId}": {
      "delete": {
        "operationId": "UsersController_removeDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device removed successfully"
          }
        },
        "summary": "Remove a registered device",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUserDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDevice"
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        },
        "summary": "Update device information",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/devices/token": {
      "patch": {
        "operationId": "UsersController_updateDeviceToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device token updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDevice"
                }
              }
            }
          }
        },
        "summary": "Update device token by previous token",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/webhooks": {
      "get": {
        "operationId": "WebhooksController_getUserWebhooks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserWebhook"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all webhooks for the authenticated user",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/v1/webhooks/{id}": {
      "get": {
        "operationId": "WebhooksController_getWebhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWebhook"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Webhook not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get webhook by ID",
        "tags": [
          "Webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "Zentik API",
    "description": "The Zentik notification API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-string"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "hasPassword": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user has a password set"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "moderator",
              "admin"
            ],
            "default": "user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "resetToken": {
            "type": "object"
          },
          "resetTokenRequestedAt": {
            "type": "object"
          },
          "emailConfirmationToken": {
            "type": "object"
          },
          "emailConfirmationTokenRequestedAt": {
            "type": "object"
          },
          "emailConfirmed": {
            "type": "boolean",
            "example": false,
            "description": "Whether the user email is confirmed"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "hasPassword",
          "role",
          "createdAt",
          "updatedAt",
          "emailConfirmed"
        ]
      },
...
      "UserDevice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "deviceToken": {
            "type": "object"
          },
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID",
              "WEB"
            ]
          },
          "deviceName": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          },
          "publicKey": {
            "type": "object",
            "description": "Public key for device-level encryption"
          },
          "privateKey": {
            "type": "object",
            "description": "Private key for device-level decryption"
          },
          "subscriptionFields": {
            "type": "object",
            "description": "Web Push subscription fields"
          },
          "onlyLocal": {
            "type": "boolean",
            "description": "Whether this device should only receive local notifications"
          },
          "lastUsed": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "id",
          "userId",
          "platform",
          "onlyLocal",
          "lastUsed",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
...
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "moderator",
              "admin"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "WebPushSubscriptionFieldsInput": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          }
        }
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "deviceToken": {
            "type": "string",
            "description": "Device token for push notifications",
            "example": "a1b2..."
          },
          "platform": {
            "type": "string",
            "description": "Platform of the device",
            "enum": [
              "IOS",
              "ANDROID",
              "WEB"
            ],
            "example": "IOS"
          },
          "deviceName": {
            "type": "string",
            "description": "Device name or identifier",
            "example": "John's iPhone"
          },
          "deviceModel": {
            "type": "string",
            "description": "Device model",
            "example": "iPhone 15 Pro"
          },
          "osVersion": {
            "type": "string",
            "description": "OS version",
            "example": "17.0.1"
          },
          "subscriptionFields": {
            "description": "Web Push subscription JSON (endpoint, p256dh, auth)",
            "allOf": [
              {
                "$ref": "#/components/schemas/WebPushSubscriptionFieldsInput"
              }
            ]
          },
          "onlyLocal": {
            "type": "boolean",
            "description": "Whether this device should only receive local notifications",
            "example": false,
            "default": false
          },
          "publicKey": {
            "type": "string",
            "description": "Public key for device-level encryption (iOS devices should provide this)",
            "example": "-----BEGIN PUBLIC KEY-----..."
          }
        },
        "required": [
          "platform"
        ]
      },
      "UpdateDeviceTokenDto": {
        "type": "object",
        "properties": {
          "oldDeviceToken": {
            "type": "string",
            "description": "Old device token to identify the device"
          },
          "newDeviceToken": {
            "type": "string",
            "description": "New device token to replace the old one"
          }
        },
        "required": [
          "oldDeviceToken",
          "newDeviceToken"
        ]
      },
      "UpdateUserDeviceDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "deviceToken": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "deviceModel": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          },
          "subscriptionFields": {
            "$ref": "#/components/schemas/WebPushSubscriptionFieldsInput"
          }
        },
        "required": [
          "deviceId"
        ]
      },
...
      "WebhookHeader": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Header key"
          },
          "value": {
            "type": "string",
            "description": "Header value"
          }
        },
        "required": [
          "key",
          "value"
        ]
      },
      "UserWebhook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ]
          },
          "url": {
            "type": "string"
          },
          "headers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookHeader"
            }
          },
          "body": {
            "type": "object",
            "description": "Webhook body content as JSON object"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "method",
          "url",
          "headers",
          "user",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}